Fix empty node-relation viz in kbn_network plugin
The version osd-1.3.2-1 of the kbn_network plugin
didn't allow to create node-relation visualizations
due to a bug in the generation of the network.
This has been fixed in the version osd-1.3.2-2.
More info: https://github.com/dlumbrer/kbn_network/issues/108

This patch bumps the version of the plugin to the one
that fixes this problem.

Signed-off-by: Santiago Dueñas <sduenas@bitergia.com>