Upgrade to OpenSearch Dashboards 2.11.1
On this new version, OSD fixes the error that
showed an empty dashboards while navigating
between them.

Due to OSD 2.11.0 removed Angular support, the
plugin 'enhanced-table' wasn't compatible anymore.
The author of this plugin was created a new version
but it's on alpha stages that we include within
this image.

Signed-off-by: Santiago Dueñas <sduenas@bitergia.com>